rest complete but still working on patch for partial update part

+++ ecommerce/views.py
@@ -5,51 +5,79 @@ from .models import User
 from .serializers import UserSerializer
 from rest_framework.views import APIView
 from django.shortcuts import get_object_or_404
+from rest_framework import status,generics,parsers
 
-class UserListCreateView(ListCreateAPIView):
+class UserListCreateView(generics.GenericAPIView):
     queryset = User.objects.all()
     serializer_class = UserSerializer
+    parser_classes = [parsers.MultiPartParser]
 
-    def get(self, request, *args, **kwargs):
-        # Handle GET request (List)
-        return super().list(request, *args, **kwargs)
-
-    def post(self, request, *args, **kwargs):
-        # Handle POST request (Create)
-        serializer = self.get_serializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        self.perform_create(serializer)
-        headers = self.get_success_headers(serializer.data)
-        return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
-
-class UserDetail(APIView):
-    def get_object(self, id):
-        return get_object_or_404(User, pk=id)
-
-    def get(self, request, id, format=None):
-        user = self.get_object(id)
-        serializer = UserSerializer(user)
+    def get(self, request, format=None):
+        drinks = User.objects.all()
+        serializer = UserSerializer(drinks, many=True)
         return Response(serializer.data)
-
-    def put(self, request, id, format=None):
-        user = self.get_object(id)
-        serializer = UserSerializer(user, data=request.data)
+    
+    def post(self, request, format=None):
+        serializer = self.get_serializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data)
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-
-    def delete(self, request, id, format=None):
+class UserDetail(generics.GenericAPIView):
+    serializer_class = UserSerializer
+    parser_classes = [parsers.MultiPartParser]
+    def get_object(self, id):
+        try:
+            return User.objects.get(pk=id)
+        except User.DoesNotExist:
+            return None
+    
+    def get(self, request, id, format=None):
         user = self.get_object(id)
-        user.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if user:
+            serializer = UserSerializer(user)
+            return Response(serializer.data)
+        return Response(status=status.HTTP_404_NOT_FOUND)
 
-    def patch(self, request, id, format=None):
-        user = self.get_object(id)
-        serializer = UserSerializer(user, data=request.data, partial=True)
+    def put(self, request, id, format=None):
+        user = get_object_or_404(User, pk=id)
+        serializer = UserSerializer(user, data=request.data)
+        
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST) 
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    
+    def delete(self, request, id, format=None):
+        user = self.get_object(id)
+        if user:
+            user.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_404_NOT_FOUND)
+    
+    # def patch(self, request, id, format=None):
+    #     user = self.get_object(id)
+    #     if user:
+    #         serializer = UserSerializer(user, data=request.data, partial=True)
+    #         if serializer.is_valid():
+    #             serializer.save()
+    #             return Response(serializer.data)
+    #         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    #     return Response(status=status.HTTP_404_NOT_FOUND)
+    
+    # from django.shortcuts import get_object_or_404
+
+    def patch(self, request, id, format=None):
+        user = get_object_or_404(User, pk=id)
+        serializer = UserSerializer(user, data=request.data, partial=True)
+        
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
